Refactored digits file into a python package and individual files for each digit
