+++ src/gameboard/digits/__init__.py
@@ -1,0 +1,11 @@
+from .digit import Digit
+from .zero import Zero
+from .one import One
+from .two import Two
+from .three import Three
+from .four import Four
+from .five import Five
+from .six import Six
+from .seven import Seven
+from .eight import Eight
+from .nine import Nine
